Set up gear workspaces
