+++ apps/gears/navigate/README.md
@@ -1,0 +1,7 @@
+# Navigate gear
+
+Placeholder
+
+## Import sequence
+
+See [ARCHITECTURES](/docs/ARCHITECTURES.md).
