+++ apps/gears/navigate/tsconfig.json
@@ -1,0 +1,3 @@
+{
+    "extends": "../tsconfig.gears.json"
+}
