+++ apps/gears/record-route/README.md
@@ -1,0 +1,7 @@
+# Record route gear
+
+Placeholder
+
+## Import sequence
+
+See [ARCHITECTURES](/docs/ARCHITECTURES.md).
