+++ apps/gears/record-route/src/index.ts
@@ -1,0 +1,1 @@
+export default {};
