+++ apps/gears/route-story/README.md
@@ -1,0 +1,7 @@
+# Route story gear
+
+Placeholder
+
+## Import sequence
+
+See [ARCHITECTURES](/docs/ARCHITECTURES.md).
