+++ apps/gears/route-story/src/index.ts
@@ -1,0 +1,2 @@
+export { RouteStoryGear } from './route-story-gear';
+export * from './model';
