+++ apps/gears/route-story/src/model.ts
