+++ apps/gears/submit-data/README.md
@@ -1,0 +1,7 @@
+# Submit data gear
+
+Placeholder
+
+## Import sequence
+
+See [ARCHITECTURES](/docs/ARCHITECTURES.md).
