+++ apps/gears/submit-data/src/index.ts
@@ -1,0 +1,1 @@
+export default {};
