+++ apps/gears/submit-data/tsconfig.json
@@ -1,0 +1,3 @@
+{
+    "extends": "../tsconfig.gears.json"
+}
