+++ apps/gears/tsconfig.gears.json
@@ -1,0 +1,26 @@
+{
+    "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "paths": {
+            "@apparatus": [
+                "../packages/apparatus/src"
+            ],
+            "@tinker-chest": [
+                "../packages/tinker-chest/src"
+            ],
+            "@ui": [
+                "../ui/src"
+            ]
+        },
+        "types": [],
+    },
+    "include": [
+        "*/src/**/*",
+        "*/test/**/*",
+        "../types/**/*"
+    ],
+    "exclude": [
+        "node_modules",
+        "*/node_modules"
+    ]
+}
