--- apps/packages/apparatus/src/gears/navigate/index.ts
@@ -1,1 +1,0 @@
-export default {};
