--- apps/packages/apparatus/src/gears/record-route/index.ts
@@ -1,1 +1,0 @@
-export default {};
