--- apps/packages/apparatus/src/gears/route-story/index.ts
@@ -1,2 +1,0 @@
-export * from './route-story';
-export * from './model';
