--- apps/packages/apparatus/src/gears/submit-data/index.ts
@@ -1,1 +1,0 @@
-export default {};
