+++ apps/packages/apparatus/src/state-warden/cartomancer/index.ts
@@ -2,3 +2,4 @@ export * from './cartomancer';
 export * from './map-layers';
 export * from './hooks';
 export * from './model';
+export * from './validation';
