+++ apps/packages/tinker-chest/src/index.ts
@@ -2,6 +2,6 @@ export * from './controls';
 export * from './data';
 export * from './preferences';
 export * from './progress';
-export * from './validation';
 export * from './parsers';
 export * from './progress';
+export * from './validation';
