+++ apps/packages/tinker-chest/src/validation/index.ts
@@ -1,1 +1,1 @@
-export * from './controls';
+export * from './validation';
