+++ apps/web/gears/route-story/package.json
@@ -11,10 +11,13 @@
     "dependencies": {
         "@the-dead-planet/nav-gauge-apparatus": "*",
         "@the-dead-planet/nav-gauge-tinker-chest": "*",
+        "@the-dead-planet/nav-gauge-gears-route-story": "*",
         "@the-dead-planet/nav-gauge-ui": "*",
-        "classnames": "^2.5.1"
+        "classnames": "^2.5.1",
+        "exif-js": "^2.3.0"
     },
     "devDependencies": {
+        "@types/exif-js": "^2.3.1",
         "maplibre-gl": "^5.16.0",
         "react": "^19.2.3",
         "react-dom": "^19.2.3"
