+++ apps/web/gears/route-story/src/index.ts
@@ -1,2 +1,1 @@
-export * from './images/image-parser';
 export * from './route-story';
