+++ apps/web/gears/route-story/src/model.ts
@@ -1,4 +1,5 @@
-import { MarkerImage, RouteTimes } from "@apparatus";
+import { MarkerImage } from "@apparatus";
+import { RouteTimes } from "@the-dead-planet/nav-gauge-gears-route-story";
 import { ParsingResultWithError } from "@tinker-chest";
 import { BehaviorSubject } from "rxjs";
